requirements.txt
flask
flask_cors
requests
google-generativeai>=0.8.3
python-dotenv>=1.0.1
gunicorn>=22.0.0
